requirements.txt
# pathway>=0.8.0 # Not supported on Windows native, use WSL or Docker
openai>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.21.0
pandas>=1.3.0
tiktoken>=0.5.0
tqdm>=4.64.0
python-dotenv>=1.0.0